requirements.txt
langchain
streamlit
streamlit_chat
faiss-cpu
sentence-transformers
huggingface_hub